+++ src/br/com/carros/validation/ValidationConstants.java
@@ -5,7 +5,7 @@ public final class ValidationConstants {
 	private ValidationConstants() {}
 	
 	public static final int TAMANHO_TEXTO = 255;
-	public static final int LIMITE_LATITUDE = 90;
-	public static final int LIMITE_LONGITUDE = 180;
+	public static final double LIMITE_LATITUDE = 90;
+	public static final double LIMITE_LONGITUDE = 180;
 
 }
